requirements.txt
numpy
Pillow
torch
torchvision
future
tensorboard
